github: add workflow to apply pull requests including git trailers
When people review pull requests or manually add Acked-by or Tested-by
trailers to comments, the information is lost and not included in the
git history. GitHub does not provide any means to include this in git
commits.

Add a workflow that can be triggered by posting a simple comment that
contains the `/apply` string. When running that workflow, checkout the
pull request code and rebase it to the latest main branch amending every
commit with git trailers.

For every "approved" GitHub review on the pull request, a `Reviewed-by`
trailer is added using the GitHub user full name (as reported on their
profile) and their email (either as reported on their profile or if not
available, extracted from the git history).

Also, a curated list of well-known trailers is extracted from all
regular comments of the pull request: `Acked-by`, `Tested-by`,
`Reviewed-by`, `Reported-by`. Their contents are included as trailers
without any further processing.

Once all commits have been rebased, the branch is fast-forward merged
and pushed (this requires that the user posting the comment has write
permissions). And the PR is marked as merged.

If any error occurs, a comment is posted on the PR for visibility.

Signed-off-by: Robin Jarry <rjarry@redhat.com>